app/requirements.txt
unstructured
"unstructured[csv]"
"unstructured[pdf]"
pandas
langchain_ollama
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
boto3 
google-cloud-storage 
azure-storage-blob 
gdown 
requests
rank_bm25
tabulate
streamlit
watchdog
fastapi
python-dotenv